:bug: Updated variable types to GoalPeriods

+++ OpenMovement.AxLE.Comms/Values/GoalConfig.cs
@@ -4,9 +4,9 @@ namespace OpenMovement.AxLE.Comms.Values
 {
     public struct GoalConfig
     {
-        public UInt16 GoalPeriod { get; set; }
+        public UInt32 GoalPeriod { get; set; }
 
-        public UInt16 GoalPeriodOffset { get; set; }
+        public UInt32 GoalPeriodOffset { get; set; }
 
         public UInt16 GoalThreshold { get; set; }
     }
